# Disable FIM Global queries - Fixing