--- a/.github/actions/check_files/manager_base.csv
+++ b/.github/actions/check_files/manager_base.csv
@@ -558,8 +558,6 @@ full_filename,owner_name,group_name,mode,type,prot_permissions,size_bytes,size_e
 /var/ossec/ruleset/sca/sca_distro_independent_linux.yml*,root,wazuh,640,file,-rw-r-----,,
 /var/ossec/ruleset/sca/web_vulnerabilities.yml*,root,wazuh,640,file,-rw-r-----,,
 /var/ossec/templates/vd_states_template.json,root,wazuh,440,file,-r--r-----,,
-/var/ossec/templates/wazuh-states-fim-files.json,root,wazuh,440,file,-r--r-----,,
-/var/ossec/templates/wazuh-states-fim-registries.json,root,wazuh,440,file,-r--r-----,,
 /var/ossec/templates/wazuh-states-inventory-packages.json,root,wazuh,440,file,-r--r-----,,
 /var/ossec/templates/wazuh-states-inventory-processes.json,root,wazuh,440,file,-r--r-----,,
 /var/ossec/templates/wazuh-states-inventory-system.json,root,wazuh,440,file,-r--r-----,,