--- a/packages/rpms/SPECS/wazuh-manager.spec
+++ b/packages/rpms/SPECS/wazuh-manager.spec
@@ -748,8 +748,6 @@ rm -fr %{buildroot}
 %attr(660, wazuh, wazuh) %ghost %{_localstatedir}/logs/ossec.json
 %dir %attr(440, root, wazuh) %{_localstatedir}/templates
 %attr(0440, root, wazuh) %{_localstatedir}/templates/vd_states_template.json
-%attr(0440, root, wazuh) %{_localstatedir}/templates/wazuh-states-fim-files.json
-%attr(0440, root, wazuh) %{_localstatedir}/templates/wazuh-states-fim-registries.json
 %attr(0440, root, wazuh) %{_localstatedir}/templates/wazuh-states-inventory-packages.json
 %attr(0440, root, wazuh) %{_localstatedir}/templates/wazuh-states-inventory-processes.json
 %attr(0440, root, wazuh) %{_localstatedir}/templates/wazuh-states-inventory-system.json