--- a/src/wazuh_db/wdb.h
+++ b/src/wazuh_db/wdb.h
@@ -103,7 +103,7 @@ typedef enum wdb_global_group_hash_operations_t {
 
 // Router provider variables
 extern ROUTER_PROVIDER_HANDLE router_agent_events_handle;
-extern ROUTER_PROVIDER_HANDLE router_fim_events_handle;
+// extern ROUTER_PROVIDER_HANDLE router_fim_events_handle; // DISABLED: FIM events are no longer processed.
 extern ROUTER_PROVIDER_HANDLE router_inventory_events_handle;
 
 typedef enum wdb_stmt {