… docs/ref/modules/inventory-harvester/README.md
… docs/ref/modules/inventory-harvester/architecture.md
… docs/ref/modules/inventory-harvester/description.md