# Description  
Closes #32005
 
From the previous implementation #31856 some improvements were need.

- The feature’s code is **not removed**, only disabled, leaving dead code in place.  
- `wdb` components has been updated to **skip forwarding any incoming message** to the router.  
- Corresponding FIM global query messages have been disabled.  

---

## Proposed Changes  

### FIM  
- Disabled processing of related messages.  

### Codebase  
- Feature remains in codebase as **dead code** for future reference.  

---

## Results and Evidence  

### Before change

```console
/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:615 at wdbi_query_checksum(): DEBUG: Agent '000' syscollector-processes range checksum: Time: 0.085 ms.
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:1936 at wdb_parse_syscollector(): DEBUG: DB(000) syscollector_processes Syscollector query.
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_right {"begin":"/usr/bin/captoinfo","checksum":"51e19f1a1fb57803689f5d6760c3ff7ccad70bf2","end":"/usr/bin/debconf-apt-progress","id":1758123116}
2025/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:615 at wdbi_query_checksum(): DEBUG: Agent '001' fim_file range checksum: Time: 0.042 ms.
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_left {"begin":"/etc/systemd/system/timers.target.wants/motd-news.timer","checksum":"2b0fa16e14a8eb4d857158c2470a5aa5e31274e0","end":"/usr/bin/apt-cdrom","id":1758123116,"tail":"/usr/bin/apt-config"}
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_right {"begin":"/usr/bin/apt-config","checksum":"28283714cafe0a13dc5850f5236e8ece6e3481d3","end":"/usr/bin/c_rehash","id":1758123116}
2025/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:615 at wdbi_query_checksum(): DEBUG: Agent '001' fim_file range checksum: Time: 0.018 ms.
2025/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:615 at wdbi_query_checksum(): DEBUG: Agent '001' fim_file range checksum: Time: 0.022 ms.
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_left {"begin":"/etc/systemd/system/timers.target.wants/motd-news.timer","checksum":"37f9588077b12b62ce16adfbcc0c28de6907e9f7","end":"/etc/xattr.conf","id":1758123116,"tail":"/sbin"}
2025/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:615 at wdbi_query_checksum(): DEBUG: Agent '001' fim_file range checksum: Time: 0.023 ms.
2025/09/17 15:31:57 wazuh-db[70977] wdb_integrity.c:177 at wdbi_report_removed(): ERROR: Invalid handle to send delete message. Agent 001
2025/09/17 15:31:57 wazuh-modulesd:syscollector[71163] logging_helper.c:40 at taggedLogFunction(): DEBUG: Sync sent: {"component":"syscollector_processes","data":{"begin":"68543","checksum":"d88c5ff7d6b3aaf41c18bc2c644f879c0bd84a06","end":"68543","id":1758123109},"type":"integrity_check_right"}
2025/09/17 15:31:57 wazuh-modulesd:syscollector[71163] logging_helper.c:40 at taggedLogFunction(): DEBUG: Sync sent: {"component":"syscollector_processes","data":{"begin":"48617","checksum":"e556201bd971aca894b1616f0118597e7e7de5d7","end":"48617","id":1758123109,"tail":"49396"},"type":"integrity_check_left"}
2025/09/17 15:31:57 wazuh-modulesd:syscollector[71163] logging_helper.c:40 at taggedLogFunction(): DEBUG: Sync sent: {"component":"syscollector_processes","data":{"begin":"67338","checksum":"e97287bc3b357ae5010185d7197e35dff30d5fe7","end":"67338","id":1758123109,"tail":"67343"},"type":"integrity_check_left"}
2025/09/17 15:31:57 wazuh-modulesd:syscollector[71163] logging_helper.c:40 at taggedLogFunction(): DEBUG: Sync sent: {"component":"syscollector_processes","data":{"attributes":{"argvs":null,"checksum":"81f2783d92282fe1313128dab6b226f5c5363f61","cmd":"/usr/libexec/upowerd","egroup":"root","euser":"root","fgroup":"root","name":"upowerd","nice":0,"nlwp":4,"pgrp":49407,"pid":"49407","ppid":1,"priority":20,"processor":2,"resident":7424,"rgroup":"root","ruser":"root","scan_time":"2025/09/17 15:31:57","session":49407,"sgroup":"root","share":1440,"size":78619,"start_time":1758108103,"state":"S","stime":191,"suser":"root","tgid":49407,"tty":0,"utime":125,"vm_size":314476},"index":"49407","timestamp":""},"type":"state"}
2025/09/17 15:31:57 wazuh-db[70977] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 000 query: syscollector_processes integrity_check_right {"begin":"68543","checksum":"d88c5ff7d6b3aaf41c18bc2c644f879c0bd84a06","end":"68543","id":1758123109}

```

### After change

```console
  root@deploy:/home/deploy/wazuh# grep AFile.txt /var/ossec/logs/ossec.log
2025/09/17 15:55:31 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_left {"begin":"/etc/terminfo/README","checksum":"683902a5aaa94e283043428da72cb662ed084f88","end":"/home/AFile.txt","id":1758124530,"tail":"/sbin"}
2025/09/17 15:55:31 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_right {"begin":"/etc/update-motd.d/50-motd-news","checksum":"9317b32578a7ec86b20106783702f861ca5e7252","end":"/home/AFile.txt","id":1758124530}
2025/09/17 15:55:31 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_right {"begin":"/etc/xattr.conf","checksum":"055a79f47de2e60e557c614313bbc85d529b6c93","end":"/home/AFile.txt","id":1758124530}
2025/09/17 15:55:31 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_left {"begin":"/etc/xattr.conf","checksum":"cb709ed6e7c1ec2a915374badae0abd9785c9358","end":"/etc/xattr.conf","id":1758124530,"tail":"/home/AFile.txt"}
2025/09/17 15:55:32 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file integrity_check_right {"begin":"/home/AFile.txt","checksum":"13a32fbf85a85e3eeee2a262b85a5b879c45a981","end":"/home/AFile.txt","id":1758124530}
2025/09/17 15:55:32 wazuh-db[80890] wdb_parser.c:310 at wdb_parse(): DEBUG: Agent 001 query: fim_file save2 {"attributes":{"attributes":"","checksum":"2e2c0495948a80362a3547bdec90721a4160c237","gid":"0","group_name":"root","hash_md5":"d41d8cd98f00b204e9800998ecf8427e","hash_sha1":"da39a3ee5e6b4b0d3255bfef95601890afd80709","hash_sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","inode":31232720,"mtime":1758124523,"perm":"rw-r--r--","size":0,"type":"file","uid":"0","user_name":"root"},"index":"/home/AFile.txt","timestamp":1758124529}
root@deploy:/home/deploy/wazuh# sqlite3 -json /var/ossec/queue/db/001.db "SELECT * FROM fim_entry;" | jq | grep AFile.txt
root@deploy:/home/deploy/wazuh# grep ERROR /var/ossec/logs/ossec.log
```

---

## Tests and Validation  

### Automated Tests  
- [ ] Unit tests updated/added where necessary.  
- [ ] Integration tests reviewed for feature disablement.  

### Manual Tests  
- [x] Compilation without warnings (Linux).  
- [ ] Memory checks (Valgrind / ASAN).  
- [x] Runtime validation: confirmed no router forwarding of FIM Global Queries.  

---

## Artifacts Affected  
- [x] wazuh_modules.  
- [ ] Configurations.  
- [ ] Documentation.  

---

## Backward Compatibility  
- [ ] No breaking changes for users relying on standard FIM.  
- [x] Global Queries feature is **disabled** (dead code remains).  

---

## Addiotional Notes  
- No configuration changes required.  
- No migration steps required.  
- Ensure upward merge plan to 4.14.0 handles conflict with remoted forwarding.  

---

## Review Checklist  
- [ ] Code reviewed (correctness, clarity).  
- [ ] Security implications reviewed.  
- [ ] Performance validated.  
- [ ] Documentation updated (marking feature disabled).  
- [ ] CI/CD pipeline green.