@GabrielEValenzuela 2025-09-17
I think index templates should keep. The logic is `dead code`, and indices can be required for later propose